move SearchActivity, SearchAdapter into ui/search that created for MVP
