+++ 2005/developer-sungjae/ArchitectureStudy/app/src/main/java/com/project/architecturestudy/ui/search/SearchAdapter.kt
@@ -1,4 +1,4 @@
-package com.project.architecturestudy.adapters
+package com.project.architecturestudy.ui.search
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
